auth/end-user/console/requirements.txt
google-cloud-bigquery[pandas]==1.7.0
google-auth-oauthlib==0.2.0
